Crate tree_sitter_html

This crate provides HTML language support for the tree-sitter parsing library.

Typically, you will use the LANGUAGE constant to add this language to a tree-sitter Parser, and then use the parser to parse some code:

let code = r#"
"#;
let mut parser = tree_sitter::Parser::new();
let language = tree_sitter_html::LANGUAGE;
parser
    .set_language(&language.into())
    .expect("Error loading HTML parser");
let tree = parser.parse(code, None).unwrap();
assert!(!tree.root_node().has_error());

Constants

HIGHLIGHTS_QUERY

The syntax highlighting query for this language.

INJECTIONS_QUERY

The injection query for this language.

LANGUAGE

The tree-sitter LanguageFn for this grammar.

NODE_TYPES

The content of the node-types.json file for this grammar.